Stop using h-prefixed properties in remark-insert-page-number
