+++ packages/remark-insert-page-number/package.json
@@ -8,6 +8,7 @@
     "url": "https://github.com/izumin5210"
   },
   "dependencies": {
+    "deepcopy": "^0.6.3",
     "defaults": "^1.0.3",
     "mdast-util-to-hast": "^2.4.0",
     "unist-builder": "^1.0.2",
