+++ packages/remark-insert-page-number/src/createHandler.js
@@ -1,0 +1,10 @@
+// @flow
+import all from 'mdast-util-to-hast/lib/all'
+
+const createHandler = (typeName: string) => (
+  function pageNumber (h: any, node: any) {
+    return h(node, typeName, node.data, all(h, node))
+  }
+)
+
+export default createHandler
